refactor(*): restructure filesystem
Closes #24
