+++ src/components/AnimatedSwitch.js
