+++ src/components/App/App.css
@@ -1,5 +1,5 @@
 .app {
-    overflow: hidden;
+  overflow: hidden;
 }
 
 .slide-enter,
