+++ src/components/App/App.js
@@ -1,7 +1,7 @@
 import React from "react";
 import "./App.css";
 import { BrowserRouter } from "react-router-dom";
-import AnimatedSwitch from "./AnimatedSwitch";
+import AnimatedSwitch from "../AnimatedSwitch";
 
 function App() {
   return (
