+++ src/components/pages/about.js
@@ -1,6 +1,6 @@
 import React from "react";
 import PropTypes from "prop-types";
-import Nav from "../Nav";
+import Nav from "../Nav/Nav";
 import "./page.css";
 
 export default function About({ classInjection }) {
