+++ src/components/pages/home.js
@@ -1,6 +1,6 @@
 import React from "react";
 import PropTypes from "prop-types";
-import Nav from "../Nav";
+import Nav from "../Nav/Nav";
 import "./page.css";
 import "./home.css";
