--- src/pages/page.css
@@ -1,17 +1,0 @@
-.page {
-    min-height: 100vh;
-    padding: 3vmin;
-    box-sizing: border-box;
-}
-
-.page--white {
-    background-color: white;
-    color: black;
-
-}
-
-.page--black {
-    background-color: black;
-    color: white;
-}
-
